feat: defaults to new favourite when unfavouriting a selected favourite in the search component
